Cleaned up components and added popup
